[#62] rename ssl module and ssl script
Because of the way Python scripts and modules get imported preferentially
the same directory in which the importing script or module resides, the
two files named ssl.py were conflicting with access to the system ssl module.
This involved:

   - urllib's access to the system ssl module when imported from within
     the irods_testing_environment (modules) subdirectory. This prevented
     the mysql odbc driver from downloading to the docker host.

   - similar problems cropped up when performing a pip install of the
     prerequisite docker-compose and GitPython modules.

As a consequence, both ssl.py files have been renamed so as not to pollute
the module namespace.  Also, we've corrected a couple of errors that were
preventing the top-level ssl.py script (now called configure_ssl.py) from
working properly.
