+++ run_topology_tests.py
@@ -11,7 +11,7 @@ from irods_testing_environment import install
 from irods_testing_environment import irods_config
 from irods_testing_environment import irods_setup
 from irods_testing_environment import services
-from irods_testing_environment import ssl
+from irods_testing_environment import ssl_setup
 from irods_testing_environment import test_utils
 
 if __name__ == "__main__":
@@ -117,7 +117,7 @@ if __name__ == "__main__":
         options.extend(['--hostnames', icat_hostname, hostname_1, hostname_2, hostname_3])
 
         if args.do_setup and args.use_ssl:
-            ssl.configure_ssl_in_zone(ctx.docker_client, ctx.compose_project)
+            ssl_setup.configure_ssl_in_zone(ctx.docker_client, ctx.compose_project)
             options.append('--use_ssl')
 
         if args.tests:
